requirements.txt
google-generativeai
streamlit
python-dotenv
youtube-transcript-api
pathlib